requirements.txt
selenium
pushbullet.py
pyyaml
